update seed file add todo edit fle

+++ db/migrate/20181105064815_create_todos.rb
@@ -4,7 +4,7 @@ class CreateTodos < ActiveRecord::Migration[5.1]
       t.integer :nominal_priority
     	t.integer :user_id
     	t.string :description
-    	t.integer :allotted_time_in_hours
+    	t.float :allotted_time_in_hours
     	t.string :dependencies
       t.timestamps
     end
